+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaReconciler.java
@@ -186,7 +186,7 @@ public class KafkaReconciler {
         this.reconciliation = reconciliation;
         this.vertx = vertx;
         this.operationTimeoutMs = config.getOperationTimeoutMs();
-        this.kafka = KafkaCluster.fromCrd(reconciliation, kafkaCr, config.versions(), oldStorage, currentReplicas);
+        this.kafka = KafkaCluster.fromCrd(reconciliation, kafkaCr, config.versions(), oldStorage, currentReplicas, config.featureGates().useKRaftEnabled());
 
         // We set the user-configured inter.broker.protocol.version if needed (when not set by the user)
         if (versionChange.interBrokerProtocolVersion() != null) {
@@ -197,6 +197,7 @@ public class KafkaReconciler {
         if (versionChange.logMessageFormatVersion() != null) {
             this.kafka.setLogMessageFormatVersion(versionChange.logMessageFormatVersion());
         }
+
         this.oldStorage = oldStorage;
         this.currentReplicas = currentReplicas;
         this.clusterCa = clusterCa;