+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.29.0
 
-* Add support for Apache Kafka 3.0.1
+* Add support for Apache Kafka 3.0.1, 3.1.1 and 3.2.0
 * Increase the size of the `/tmp` volumes to 5Mi to allow unpacking of compression libraries
 * Use `/healthz` endpoint for Kafka Exporter health checks
 * Renew user certificates in User Operator only during maintenance windows
@@ -18,6 +18,8 @@
   * `full`: this mode runs a full rebalance moving replicas across all the brokers in the cluster. This is the default one if not specified.
   * `add-brokers`: after scaling up the cluster, this mode is used to move replicas to the newly added brokers specified in the custom resource.
   * `remove-brokers`: this mode is used to move replicas off the brokers that are going to be removed, before scaling down the cluster.
+* **Experimental** KRaft mode (ZooKeeper-less Kafka) which can be enabled using the `UseKRaft` feature gate.
+  **Important: Use it for development and testing only!**
 
 ### Changes, deprecations and removals